refactored the code, made App.js smaller and divided it into separate components
